// src/lib/supabase/types/database.ts
export type Database = {
  public: {
    Tables: {
// ...
      products: {
        Row: {
          brand: string | null
          category_id: string | null
          created_at: string
          description: string | null
          id: string
          image_url: string | null
          name: string
          pao: number | null
          subcategory_id: number | null
        }
        Insert: {
          brand?: string | null
          category_id?: string | null
          created_at?: string
          description?: string | null
          id?: string
          image_url?: string | null
          name: string
          pao?: number | null
          subcategory_id?: number | null
        }
        Update: {
          brand?: string | null
          category_id?: string | null
          created_at?: string
          description?: string | null
          id?: string
          image_url?: string | null
          name?: string
          pao?: number | null
          subcategory_id?: number | null
        }
        Relationships: [
          {
            foreignKeyName: "products_category_id_fkey"
            columns: ["category_id"]
            isOneToOne: false
            referencedRelation: "product_category"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "products_subcategory_id_fkey"
            columns: ["subcategory_id"]
            isOneToOne: false
            referencedRelation: "product_subcategory"
            referencedColumns: ["id"]
          },
        ]
      }
      user_products: {
        Row: {
          batch_code: string | null
          created_at: string
          expiry_date: string | null
          id: string
          image_url: string | null
          is_favorite: boolean | null
          last_used_at: string | null
          notes: string | null
          open_date: string | null
          product_id: string | null
          purchase_date: string | null
          status: number | null
          updated_at: string
          user_id: string | null
        }
        Insert: {
          batch_code?: string | null
          created_at?: string
          expiry_date?: string | null
          id?: string
          image_url?: string | null
          is_favorite?: boolean | null
          last_used_at?: string | null
          notes?: string | null
          open_date?: string | null
          product_id?: string | null
          purchase_date?: string | null
          status?: number | null
          updated_at?: string
          user_id?: string | null
        }
        Update: {
          batch_code?: string | null
          created_at?: string
          expiry_date?: string | null
          id?: string
          image_url?: string | null
          is_favorite?: boolean | null
          last_used_at?: string | null
          notes?: string | null
          open_date?: string | null
          product_id?: string | null
          purchase_date?: string | null
          status?: number | null
          updated_at?: string
          user_id?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "user_products_product_id_fkey"
            columns: ["product_id"]
            isOneToOne: false
            referencedRelation: "products"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "user_products_status_fkey"
            columns: ["status"]
            isOneToOne: false
            referencedRelation: "product_status"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "user_products_user_id_fkey"
            columns: ["user_id"]
            isOneToOne: false
            referencedRelation: "users"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}